requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.23,<1.0
supabase>=2.6,<3.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<1.0
pytest>=7,<9.0
openai>=1.3,<2.0
pillow>=10,<11
python-multipart>=0.0.9
aiofiles>=23,<24
